Apply default theme on tab without container
Fallback on a builtin light theme if the system theme is selected.
We currently also lost some colors and properties for the Firefox
dark/light builtin theme with this method because the API doesn't
let us defines colors and properties containing for example
color-mix() function like the Firefox builtin theme can.